+++ CHANGELOG.md
@@ -1,1 +1,12 @@
 # Changelog
+
+## [Version 0.1.0] - Initial Release - 2021-10
+
+* GLM regression in the doctor
+* GLM binary classification in the doctor
+* Elastic Net Penalization
+* Sample Weighting
+* Offset and Exposure Support
+* Spline Regression as a recipe
+* Spline Regression as a prepare step
+* Actual vs Predicted view in Analysis and Saved Models